Move log and config_data into server classes library
We originally had the log and the server config object in the main
server directory, but it doesn't really make a whole lot of sense
there.  Doing any sort of testing or linking of things becomes a
rather large hassle.

We'll move both objects into the server classes library, so we can
just link with the lib and be done with it.

Unit test suite has some changes, but they will be committed at
another time.
