+++ server/classes/log.cc
